requirements.txt
pulsar-client
findspark